requirements.txt
Django==3.1
Celery
redis
django-celery-results
djangorestframework
requests